requirements.txt
tensorflow==2.9.1
keras==2.9.0
Pillow==7.1.2
Shapely==1.8.5.post1
opencv-python==4.6.0.66
pandas==1.3.5
numpy==1.21.6
matplotlib==3.2.2